fix(contract-deploy): dependency cycle
