+++ packages/sdk/src/modules/transfers/BaseTransfer.ts
@@ -1,12 +1,4 @@
-import {
-  BN,
-  hexlify,
-  Provider,
-  ScriptTransactionRequest,
-  TransactionRequest,
-  transactionRequestify,
-  TransactionResponse,
-} from 'fuels';
+import { hexlify, TransactionRequest, TransactionResponse } from 'fuels';
 import {
   ITransaction,
   ITransactionResume,
@@ -15,8 +7,6 @@ import {
 } from '../../api';
 import { Vault } from '../vault/Vault';
 import { delay } from '../../../test/utils';
-import { FAKE_WITNESSES } from './helpers';
-import { ITransferEstimateFee } from './types';
 
 export interface BaseTransferLike<T extends TransactionRequest> {
   name: string;
@@ -165,32 +155,5 @@ export class BaseTransfer<T extends TransactionRequest> {
       status: transaction.status,
     };
     return result;
-  }
-
-  public static async estimateFee(
-    transaction: TransactionRequest,
-    provider: Provider,
-    required_witnesses: number,
-  ): Promise<ITransferEstimateFee> {
-    try {
-      const _tx = transactionRequestify(transaction);
-
-      _tx.witnesses = Array.from(
-        { length: required_witnesses },
-        () => FAKE_WITNESSES,
-      );
-
-      const fee = await provider.estimateTxGasAndFee({
-        transactionRequest: transactionRequestify(_tx),
-      });
-
-      return {
-        ...fee,
-        bako_max_fee: fee.maxFee.add(fee.minFee),
-        bako_gas_limit: fee.gasLimit,
-      };
-    } catch (e) {
-      return Promise.reject(e);
-    }
   }
 }
