+++ packages/sdk/src/modules/transfers/Transfer.ts
@@ -1,19 +1,14 @@
+import { BN, Provider, TransactionRequest, transactionRequestify } from 'fuels';
 
-import { TransactionRequest, } from 'fuels';
+import { ECreationTransactiontype, TransferFactory } from './types';
 
-import {
-  ECreationTransactiontype,
-  TransferFactory,
-} from './types';
-
-import { identifyCreateTransactionParams } from './helpers';
+import { FAKE_WITNESSES, identifyCreateTransactionParams } from './helpers';
 import { BaseTransfer } from './BaseTransfer';
-
 
 /**
  * `Transfer` are extension of ScriptTransactionRequest, to create and send transactions
  */
-export class Transfer extends BaseTransfer<TransactionRequest>{
+export class Transfer extends BaseTransfer<TransactionRequest> {
   /**
    * Create a new transaction instance
    *
@@ -25,7 +20,6 @@ export class Transfer extends BaseTransfer<TransactionRequest>{
    * @returns return a new Transfer instance
    */
   public static async instance(param: TransferFactory): Promise<Transfer> {
-
     const { payload, type } = await identifyCreateTransactionParams(param);
 
     if (!(type in ECreationTransactiontype)) {
@@ -33,5 +27,41 @@ export class Transfer extends BaseTransfer<TransactionRequest>{
     }
 
     return new Transfer(payload);
+  }
+
+  public static async estimateFee(
+    transaction: TransactionRequest,
+    provider: Provider,
+    required_witnesses: number,
+  ): Promise<{
+    minGas: BN;
+    minFee: BN;
+    maxGas: BN;
+    maxFee: BN;
+    gasPrice: BN;
+    gasLimit: BN;
+    bako_max_fee: BN;
+    bako_gas_limit: BN;
+  }> {
+    try {
+      const _tx = transactionRequestify(transaction);
+
+      _tx.witnesses = Array.from(
+        { length: required_witnesses },
+        () => FAKE_WITNESSES,
+      );
+
+      const fee = await provider.estimateTxGasAndFee({
+        transactionRequest: transactionRequestify(_tx),
+      });
+
+      return {
+        ...fee,
+        bako_max_fee: fee.maxFee.add(fee.minFee),
+        bako_gas_limit: fee.gasLimit,
+      };
+    } catch (e) {
+      return Promise.reject(e);
+    }
   }
 }
